+++ tests/testthat/test-i_html.R
@@ -99,6 +99,11 @@ test_that("Interactive tables won't fail when using different options", {
     cols_merge(columns = c(low, high), pattern = "{1}&mdash;{2}") %>%
     cols_label(open = "open/close", low = "low/high") %>%
     opt_interactive()
+  # #1759 sub_ works
+  tbl_gt_i_29 <- exibble %>%
+    gt() %>%
+    sub_missing(rows = 1:7) %>%
+    opt_interactive()
 
   capture_output(expect_no_error(tbl_gt_i_01))
   capture_output(expect_no_error(tbl_gt_i_02))
@@ -128,5 +133,6 @@ test_that("Interactive tables won't fail when using different options", {
   capture_output(expect_no_error(tbl_gt_i_26))
   capture_output(expect_no_error(tbl_gt_i_27))
   capture_output(expect_no_error(tbl_gt_i_28))
+  capture_output(expect_no_error(tbl_gt_i_29))
 
 })